Add `--check` option for `veryl build`
