+++ crates/veryl/src/cmd_fmt.rs
@@ -1,13 +1,10 @@
 use crate::OptFmt;
-use console::{Style, style};
+use crate::diff::print_diff;
 use log::{debug, info};
 use miette::{IntoDiagnostic, Result, WrapErr};
-use similar::{ChangeTag, TextDiff};
-use std::fmt;
 use std::fs;
 use std::fs::OpenOptions;
 use std::io::Write;
-use std::path::Path;
 use veryl_analyzer::Analyzer;
 use veryl_formatter::Formatter;
 use veryl_metadata::Metadata;
@@ -62,54 +59,5 @@ impl CmdFmt {
         }
 
         Ok(all_pass)
-    }
-}
-
-struct Line(Option<usize>);
-
-impl fmt::Display for Line {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        match self.0 {
-            None => write!(f, "    "),
-
-            Some(idx) => write!(f, "{:<4}", idx + 1),
-        }
-    }
-}
-
-fn print_diff(file: &Path, org: &str, new: &str) {
-    let diff = TextDiff::from_lines(org, new);
-
-    println!("Diff in {}", file.to_string_lossy());
-
-    for (idx, group) in diff.grouped_ops(3).iter().enumerate() {
-        if idx > 0 {
-            println!("{:-^1$}", "-", 80);
-        }
-        for op in group {
-            for change in diff.iter_inline_changes(op) {
-                let (sign, s) = match change.tag() {
-                    ChangeTag::Delete => ("-", Style::new().red()),
-                    ChangeTag::Insert => ("+", Style::new().green()),
-                    ChangeTag::Equal => (" ", Style::new().dim()),
-                };
-                print!(
-                    "{}{} |{}",
-                    style(Line(change.old_index())).dim(),
-                    style(Line(change.new_index())).dim(),
-                    s.apply_to(sign).bold(),
-                );
-                for (emphasized, value) in change.iter_strings_lossy() {
-                    if emphasized {
-                        print!("{}", s.apply_to(value).underlined().on_black());
-                    } else {
-                        print!("{}", s.apply_to(value));
-                    }
-                }
-                if change.missing_newline() {
-                    println!();
-                }
-            }
-        }
     }
 }
